inference/requirements.txt
fastapi
uvicorn
pydantic
numpy
tensorflow
joblib
pandas
requests
fastapi_jwt_auth
python-jose[cryptography]
passlib[bcrypt]